multiplier: Move product signal before register
The output does not need to be registered.  The flops are only for
accumulation purposes, but this does not matter after the last
accumulation.  Note that the NOR optimization needed to be reverted
to facilitate correct timing.  Therefore, the multiplier will have a
1 cycle delay when used continuously, but this is okay since we will
likely want to register the output product somewhere anyway.