shifter: Throttling for continuous operation
Due to the design of the shifter/rotator, there will always be an
additional clock cycle between successive operations.  The new input
value must be loaded in this cycle.  Contrary to the divider and
multiplier, the new value cannot be loaded at the time that o_finished
is asserted, since the output value is combinational, and this will
overwrite the previous answer.  Unfortunately the output must remain
combinational in order to support a shift by 0.  This is okay though,
because the ALU will most likely need a cycle to save the output value
in a register